// apps/backend/src/modules/registrations/domain/enums.ts
export enum RegistrationStatus {
  PENDENTE = 'pendente',
  EM_ANALISE = 'em_analise',
  APROVADO = 'aprovado',
  REJEITADO = 'rejeitado',
}

export enum DocumentStatus {
  PENDENTE = 'pendente',
  EM_ANALISE = 'em_analise',
  APROVADO = 'aprovado',
  REJEITADO = 'rejeitado',
}

export enum RegistrationDocumentType {
  ATESTADO_MEDICO = 'atestado_medico',
  DOCUMENTO_IDENTIDADE = 'documento_identidade',
  COMPROVANTE_PAGAMENTO = 'comprovante_pagamento',
  OUTROS = 'outros',
}